requirements.txt
bcrypt==3.2.0
cachelib==0.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Ext==0.1
Flask-Sessionstore==0.4.5
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Naked==0.1.31
peewee==3.14.4
pycparser==2.20
pycryptodome==3.10.1
PyYAML==5.4.1
requests==2.26.0
shellescape==3.8.1
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
